Create MiddleElementSinglyLinked.java
solution 1
This method works on the idea of counting the total no. of nodes in the given linked list and using this information to find out which is the middle element.


This algorithm involves the following steps:

Traverse all the nodes in the given linked list starting from the head node.

Take a variable to keep track of the count of total nodes traversed. Lets call this variable nodeCount.

Once you have the count of total nodes in a linked list, you can get the middle node using middleNode = nodeCount/2.
Now traverse the linked list again from the head node until middleNode. To do this take a variable currentNodeCount to keep track of the current node count.  As you visit each node, increment this counter by 1. When currentNodeCount is equal to middleNode we have reached the middle node, return the middle node.
*******
Time Complexity: O(n)
Since we make 2 passes (one full pass and one pass till the middle node) on the given linked list, time complexity is O(n) + O(n/2) which in general can be represented as O(n).
Space Complexity: O(1)
****************************************

Solution 2
************************************************************************************************************************
Take two pointers variables, lets call them fast and slow.
Initially both fast and slow points to the head node of the given linked list.
slow pointer moves one step at a time and fast pointer moves two steps at a time.
We need to move the fast and slow pointers from the head node until the fast pointer reaches end of the list.
When the fast pointer reaches the last node, slow pointer would have reached the middle node since it is moving at half the pace as the fast pointer.
So when the fast pointer reaches the last node, return the slow pointer node as your result.
************************************************************************************************************************